LUDOC-385 dom: fix figure to match description
The figure describes 4MB OST stripe_size, but the diagram shows five
blocks within the "4MB" range on the OSTs, with the first 1MB block
being (correctly) sparse because of the DoM component.  Fix this to
show the proper 4MB range on the OST objects.

Also, show the MB offset of each OST stripe as "4, 8, 4N" rather than
"2, 3, N" to better explain how Lustre striping works.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I1b9487c4b19cecbdc738744381166968373ebbe5
Reviewed-on: https://review.whamcloud.com/32263
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <joseph.gmitter@intel.com>