LU-6864 mdc: move RPC semaphore code to lustre/osp
The "MDC RPC semaphore" is no longer used by MDC code since patch
http://review.whamcloud.com/14374 "LU-5319 mdc: manage number of
modify RPCs in flight" landed.  It is only still used by the OSP
currently.  While there are plans to remove this from the OSP as
well, it makes sense to move all of this code from MDC to OSP so
that it will also be cleaned up when that functionality lands.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I1c4c9a45f308a29600631ac7fed40e592b3ebbe5
Reviewed-on: https://review.whamcloud.com/32412
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Lai Siyao <lai.siyao@whamcloud.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Mike Pershin <mpershin@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>