LU-11012 doc: improve the lfs-setstripe.1 man page
Restructure the lfs-setstripe.1 man page to be more like a standard
man page.  The DESCRIPTION section provides a high-level overview of
the command, and describes the various different ways that it can be
used.  The various options are described in the STRIPE_OPTIONS and
COMPONENT_OPTIONS sections.

Describe how default layouts are inherited for plain and composite
files, and that different pools can be specified for each component.

Add description of the --copy option.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I7ba3f3f4dba6825f283dd6874ae1988914c774fb
Reviewed-on: https://review.whamcloud.com/32463
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Joseph Gmitter <jgmitter@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>