LU-10447 build: fix checkpatch.pl to not mangle case
Fix the checkpatch.pl script to not mangle the case of the lines
in the spelling.txt file, so that the replacement text can be
found, and is shown as intended.

Add checks for "GETSTRIPE" and "SETSTRIPE" to spelling.txt.

Sort the spelling.txt file so that it is easier to see what lines
are present in the file.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Id3a3319bc62898fbdc94573512cd6a665b13cd11
Reviewed-on: https://review.whamcloud.com/32865
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>