LUDOC-249 migrate: Improve lfs_migrate documentation
List the new method of max_create_count=0 for disabling the MDS to
create new objects on OSTs during migration.  This avoids issues
when using "deactivate" where the OST objects were not being deleted
as objects are being moved off of the OST.

Separate the directory migration from the data migration more clearly.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ic8fa99271d8d543ec93ddac3917c8a7cad49973b
Reviewed-on: https://review.whamcloud.com/32882
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <jgmitter@whamcloud.com>