LU-11181 mdt: include DoM in inode ratio discussion
Reference Data-on-MDT in the discussion about ldiskfs inode ratios.
We default to 2KB/inode on the MDT, but with DoM we may want to
change this to 64KB/inode or higher, depending on the average size
of small files stored on the MDT.

Also mention that "lfs df -i" and "df -i" will limit the reported
available inode count to the total number of available inodes on
the OSTs.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I53c4d35129ffb5c7732771a6ddab3aa723758872
Reviewed-on: https://review.whamcloud.com/32887
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <jgmitter@whamcloud.com>