LU-10627 tests: don't use libtool wrapper for applications
It is a common pratice of lustre developers to test within the
lustre tree without actually installing lustre onto the local
node. In order for this to work the test suite needs to use
the binary executables instead of the libtool executable wrappers.
Add in the libtool LDFLAG to prevent the creation of the wrappers
for the lustre utils. Additionally properly set LD_LIBRARY_PATH
to where libtool caches the dynamic libraries.

Change-Id: I9570fcb65b927463076f28c47ecec924602bef4e
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/32835
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Oleg Drokin <green@whamcloud.com>