LU-8130 libcfs: prepare rhashtable support
Linux has a resizeable hashtable implementation in lib,
so we should use that instead of having one in libcfs.
In the process we gain lockless lookup which should be
a performance boost. All modern distributions Lustre
support has rhashtable support but a few pieces are
missing for systems running a 4.4 kernel. The other
target platforms have the full implementation we need.

Test-Parameters: trivial

Change-Id: I63d5b7dae9d52eed12dbefed8ca6062af33efd30
Signed-off-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-on: https://review.whamcloud.com/32102
Reviewed-by: John L. Hammond <jhammond@whamcloud.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>