LU-11329 utils: add get_maintainer.pl script
The get_maintainer.pl script can be given a patch file or list of
pathnames and will look at the MAINTAINERS file for different code
subsections and the developers that are maintaining those sections.
If no maintainer is found, it will also check the git commit history
and "git blame" to see which developers were active on the specified
file(s) and use them.

This patch adds a copy of the get_maintainers.pl and MAINTAINERS files
to the Lustre tree to make it easier to find reviewers for the code.

Test-Parameters: trivial
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I5f2e717227a7c86c924e7edb889e087c7c3ebbe5
Reviewed-on: https://review.whamcloud.com/33104
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Wang Shilong <wshilong@ddn.com>
Reviewed-by: John L. Hammond <jhammond@whamcloud.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>