LU-10899 osc: move obdo_cache to OSC code
The obdo_cache slab is only used by the OSC code today, so it does
not need to be allocated in obdclass on servers.  Move it to only
be allocated when the OSC module is loaded.

Rename obdo_cachep to osc_obdo_kmem to match other slab caches
created by the OSC.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I6f70adb1f5d0e76c72d797a65b413df1533ebbe5
Reviewed-on: https://review.whamcloud.com/33141
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Bobi Jam <bobijam@hotmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>