LU-11057 tests: exercise 'network' client mount option
Test 'network' client mount option that enables to limit connections
from the client to be on the specified network NID.

This option is currently not compatible with Dynamic Peer Discovery.
test_31 will have to be updated once Dynamic Peer Discovery is
modified so that it does not conflict with 'network' client mount
option.

Signed-off-by: Sebastien Buisson <sbuisson@ddn.com>
Change-Id: I13f016fad52a7b73524f652b369c6434985fc0c6
Reviewed-on: https://review.whamcloud.com/33189
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Nunez <jnunez@whamcloud.com>
Reviewed-by: Amir Shehata <ashehata@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>