LUDOC-11 misc: remove some 'l23' conditions add 'l2C'
Remove many of the old Lustre 2.3 annotations from the manual,
since there are few, if any, users working with such old versions
of Lustre, and they can always reference an older version of the
manual.

Improve content in affected sections of the manual.

Add the 'l2C' condition for the Lustre 2.12 release.  This is needed
to document the new RPC checksum types added to the 2.12 release.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I8d10c6451899a470a12600452bc905ad0b545889
Reviewed-on: https://review.whamcloud.com/33378
Tested-by: Jenkins
Reviewed-by: Joseph Gmitter <jgmitter@whamcloud.com>